reorded src dirs for docker compse build
