+++ docker-compose.yml
@@ -27,9 +27,9 @@ services:
               count: all
               capabilities: [gpu]
     # command: python3 train_cnn.py
-    # command: python3 script.py
+    command: python3 script.py
     # command: python3 main.py
-    command: python3 plt_rebuild.py
+    # command: python3 plt_rebuild.py
     # command: python3 sweeps.py
     # command: /bin/bash # interactive shell new
     stdin_open: true
