+++ srcold/helper/util.py
@@ -23,7 +23,6 @@ from sklearn.model_selection import KFold
 from icecream import ic
 from time import time as timer
 from dotenv import load_dotenv
-from configs.augmentation_config import AugmentationConfig
 
 
 def generate_model_image(model: torch.nn.Module, device:str):
@@ -105,7 +104,7 @@ class AudioDataset(Dataset):
                  augmentations_per_sample: int = 0,
                  augmentations: list[str] = [],
                  num_channels: int = 1,
-                 config: Optional[dict] = None) -> None:
+                 config: dict = None) -> Dataset:
         self.paths = data_paths
         self.feature_extractor = feature_extractor
         self.classes, self.class_to_idx = find_classes(data_path)
@@ -126,16 +125,14 @@ class AudioDataset(Dataset):
         self.target_duration = target_duration
         self.target_length = target_duration * self.target_sr
         self.augmentations_per_sample = augmentations_per_sample
-        self.config = config or {}
+        self.config = config
 
         total_samples = (augmentations_per_sample + 1) * len(self.paths)
         self.audio_tensors = torch.empty(total_samples, num_channels, self.target_sr * target_duration)
         self.class_indices = []
 
-        if self.augmentations_per_sample > 0 and augmentations:
+        if self.augmentations_per_sample > 0:
             self.composed_transform = create_augmentation_pipeline(augmentations, self.config)
-        else:
-            self.composed_transform = None
 
         # Load original audio samples
         for index, path in enumerate(self.paths):
@@ -388,8 +385,8 @@ def train_test_split_custom(
     inference_size: float = 0.1,
     seed: int = 42, 
     augmentations_per_sample: int = 3,
-    augmentations: Optional[list[str]] = None,
-    config: Optional[AugmentationConfig] = None
+    augmentations: list[str] = None,
+    config: dict = None
 ):
     def split_dataset(data, val_size, test_size, inference_size, random_state=None):
         train_size = 1.0 - (val_size + test_size + inference_size)
@@ -425,26 +422,24 @@ def train_test_split_custom(
     test_paths = [all_paths[i] for i in test_indices]
     inference_paths = [all_paths[i] for i in inference_indices]
 
-    # Create AudioDataset instances with proper config handling
-    augmentations = augmentations or []
-    config_dict = config.aug_configs if isinstance(config, AugmentationConfig) else {}
-    
+
+    # Create AudioDataset instances
     train_dataset = AudioDataset(data_path,
-                                train_paths,
-                                feature_extractor,
-                                augmentations_per_sample=augmentations_per_sample,
-                                augmentations=augmentations,
-                                config=config_dict)
+                                 train_paths,
+                                 feature_extractor,
+                                 augmentations_per_sample=augmentations_per_sample,
+                                 augmentations=augmentations,
+                                 config=config)
     
     val_dataset = AudioDataset(data_path, 
                                val_paths, 
                                feature_extractor,
                                augmentations_per_sample=augmentations_per_sample,
                                augmentations=augmentations,
-                               config=config_dict)
+                               config=config)
     
-    test_dataset = AudioDataset(data_path, test_paths, feature_extractor, config=config_dict)
-    inference_dataset = AudioDataset(data_path, inference_paths, feature_extractor, config=config_dict)
+    test_dataset = AudioDataset(data_path, test_paths, feature_extractor, config=config)
+    inference_dataset = AudioDataset(data_path, inference_paths, feature_extractor, config=config)
     
     print(f"Lengths: Train: {len(train_dataset)}, Validation: {len(val_dataset)}, "
           f"Test: {len(test_dataset)}, Inference: {len(inference_dataset)}")
@@ -492,8 +487,8 @@ def k_fold_split_custom(
     inference_size: float = 0.1,
     seed: int = 42,
     augmentations_per_sample: int = 3,
-    augmentations: Optional[list[str]] = None,
-    config: Optional[AugmentationConfig] = None
+    augmentations: list[str] = None,
+    config: dict = None
 ) -> list[tuple]:
     """
     Creates k-fold splits of the dataset for cross validation.
@@ -535,14 +530,14 @@ def k_fold_split_custom(
             feature_extractor,
             augmentations_per_sample=augmentations_per_sample,
             augmentations=augmentations,
-            config=config.aug_configs if isinstance(config, AugmentationConfig) else {}
+            config=config
         )
         
         val_dataset = AudioDataset(
             data_path,
             fold_val_paths,
             feature_extractor,
-            config=config.aug_configs if isinstance(config, AugmentationConfig) else {}
+            config=config
         )
         
         fold_datasets.append((train_dataset, val_dataset))
@@ -550,15 +545,11 @@ def k_fold_split_custom(
     
     # Create inference dataset
     inference_paths = [all_paths[i] for i in inference_indices]
-    inference_dataset = AudioDataset(data_path, inference_paths, feature_extractor, config=config.aug_configs if isinstance(config, AugmentationConfig) else {})
+    inference_dataset = AudioDataset(data_path, inference_paths, feature_extractor, config=config)
     
     end_time = timer()  # End timer for dataset initialization
     dataset_init_time = end_time - start_time
     print(f"Dataset initialization took {dataset_init_time:.2f} seconds")
-    
-    # Handle config properly
-    augmentations = augmentations or []
-    config_dict = config.aug_configs if isinstance(config, AugmentationConfig) else {}
     
     return fold_datasets, inference_dataset
